+++ one/remote/globus.py
@@ -37,6 +37,7 @@ Path('/path/to/new/location/path/to/file.ext')
 import os
 import re
 import sys
+import asyncio
 import logging
 from uuid import UUID
 from pathlib import Path, PurePosixPath, PurePath, PureWindowsPath
@@ -45,7 +46,7 @@ from functools import partial
 
 import globus_sdk
 from globus_sdk import TransferAPIError, GlobusAPIError, NetworkError, GlobusTimeoutError, \
-    GlobusConnectionError, GlobusConnectionTimeoutError
+    GlobusConnectionError, GlobusConnectionTimeoutError, GlobusSDKUsageError
 from iblutil.io import params as iopar
 
 from one.alf.spec import is_uuid
@@ -786,6 +787,9 @@ class Globus(DownloadClient):
             If True, remove the UUID from the returned filenames.
         return_size : bool
             If True, return the size of each listed file in bytes.
+        max_retries : int
+            The number of times to retry the remote operation before raising. Increasing this may
+            mitigate unstable network issues.
 
         Returns
         -------
@@ -830,7 +834,7 @@ class Globus(DownloadClient):
         target_endpoint : uuid.UUID, str
             The Globus destination endpoint. May be a UUID or a key in the Globus.endpoints
             attribute.
-        source_paths : list of str, Path or PurePath
+        source_paths : list of str, pathlib.Path or pathlib.PurePath
             The absolute or relative Globus paths of source files to moves.  Note: if endpoint is
             a UUID, the path must be absolute.
         target_paths : list of str, Path or PurePath
@@ -838,7 +842,7 @@ class Globus(DownloadClient):
             is a UUID, the path must be absolute.
         timeout : int
             Maximum time in seconds to wait for the task to complete.
-        kwargs
+        **kwargs
             Optional arguments for globus_sdk.TransferData.
 
         Returns
@@ -933,68 +937,41 @@ class Globus(DownloadClient):
                 _logger.warning('\nGlobus experienced a network error, retrying.')
                 self.run_task(globus_func, retries=(retries - 1), timeout=timeout)
 
-    def delete_dataset(self, dataset, alyx=None):
+    async def task_wait_async(self, task_id, polling_interval=10, timeout=10):
         """
-        Delete a dataset off Alyx and remove file record from all Globus repositories
+        Asynchronously wait until a Task is complete or fails, with a time limit.
+
+        If the task status is ACTIVE after timout, returns False, otherwise returns True.
 
         Parameters
         ----------
-        dataset : uuid.UUID, str, dict
-            The dataset record or ID to delete.
-        alyx : one.webclient.AlyxClient
-            An Alyx instance from which to delete the dataset.
+        task_id : str, uuid.UUID
+            A Globus task UUID to wait on for completion.
+        polling_interval : float
+            Number of seconds between queries to Globus about the task status. Minimum 1 second.
+        timeout : float
+            Number of seconds to wait in total. Minimum 1 second.
 
         Returns
         -------
-        list of uuid.UUID
-            A list of Globus delete task IDs.
+        bool
+            True if status not ACTIVE before timeout. False if status still ACTIVE at timeout.
         """
-        from itertools import groupby, starmap
-        from collections import defaultdict
-        from one.alf.files import add_uuid_string
-
-        alyx = alyx or AlyxClient()
-        if is_uuid(dataset):
-            did = dataset
-            dataset = alyx.rest('datasets', 'read', id=did)
-        else:
-            did = dataset['url'].split('/')[-1]
-
-        files_by_repo = defaultdict(list)  # uuid.UUID -> [pathlib.PurePosixPath]
-        file_records = filter(lambda x: x['exists'], dataset['file_records'])
-        for repo, record in groupby(file_records, lambda x: x['data_repository']):
-            if not record['globus_id']:
-                raise NotImplementedError
-            if repo not in self.endpoints:
-                self.add_endpoint(repo, alyx=alyx)
-            filepath = PurePosixPath(record['relative_path'])
-            if 'flatiron' in repo:
-                filepath = add_uuid_string(filepath, did)
-            files_by_repo[repo].append(filepath)
-
-        # Delete the files
-        task_ids = list(starmap(self.delete_data, files_by_repo.items()))
-        # Delete the dataset from Alyx
-        alyx.rest('datasets', 'delete', id=did)
-        return task_ids
-
-    async def task_wait_async(self, task_id, polling_interval=10):
-        import asyncio
         if polling_interval < 1:
-            raise ValueError
-        TIMEOUT = 60 * 60
-        polling_interval = min(TIMEOUT, polling_interval)
+            raise GlobusSDKUsageError('polling_interval must be at least 1 second')
+        if timeout < 1:
+            raise GlobusSDKUsageError('timout must be at least 1 second')
+        polling_interval = min(timeout, polling_interval)
         waited_time = 0
         while True:
             task = self.client.get_task(task_id)
             status = task['status']
-            if status != "ACTIVE":
+            if status != 'ACTIVE':
                 return True
 
-            # make sure to check if we timed out before sleeping again, so we
-            # don't sleep an extra polling_interval
+            # check if we timed out before sleeping again
             waited_time += polling_interval
-            if waited_time >= TIMEOUT:
+            if waited_time >= timeout:
                 return False
 
             await asyncio.sleep(polling_interval)