+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.2.2]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.3.0]
+
+### Added
+
+- Globus.delete_data method for deleting files and folders from an endpoint
+- Globus.task_wait_async is an asynchronous implementation of the TransferClient.task_wait method
+
+## [2.2.2]
 
 ### Modified